Move layout-v2 GEMM2 sources into kernels/mxmoe_g2 and update call sites.
