+++ aiter/ops/flydsl/kernels/mxmoe_g2/__init__.py
@@ -1,0 +1,15 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (C) 2025-2026 FlyDSL Project Contributors
+"""Layout-v2 MXFP4 MoE GEMM2 (dispatcher + atoms / K-loop / epilog / scheduler)."""
+
+from .mxmoe_dispatcher import (
+    _validate_v2_gemm2_dtypes,
+    compile_gemm2_a4w4_port,
+    mxfp4_moe_gemm2,
+)
+
+__all__ = [
+    "compile_gemm2_a4w4_port",
+    "mxfp4_moe_gemm2",
+    "_validate_v2_gemm2_dtypes",
+]
