+++ aiter/ops/flydsl/kernels/mxmoe_g2/mxmoe_g2_epilog.py
@@ -17,8 +17,8 @@ from flydsl.expr.typing import (
 from flydsl.expr.typing import Vector as Vec
 from flydsl.expr.typing import as_ir_value as _raw
 
-from .mxfp4_gemm_common import _fabs_f32 as fabs_f32
-from .mxfp4_gemm_common import (
+from ..mxfp4_gemm_common import _fabs_f32 as fabs_f32
+from ..mxfp4_gemm_common import (
     _inline_dpp_pair_amax,
     _inline_dpp_quad_amax,
     _udiv,
